// Web/RMA-Member-Portal/memberportal-app/src/app/shared/enums/policy-status.enum.ts
export enum PolicyStatus {
    Active = 1,
    Cancelled = 2,
    Transferred = 3,
    Expired = 4,
    Lapsed = 5,
    Legal = 6,
    Paused = 7,
    PendingFirstPremium = 8,
    PreLegal = 9,
    PendingCancelled = 10,
    PendingReinstatement = 11,
    PendingContinuation = 12,
    NotTakenUp = 13,
    Continued = 14,
    Reinstated = 15,
    PremiumWaivered = 16,
    FreeCover = 20
  }
  
  